meet: move assistant/src/tools/meet → skills/meet-join/tools
