+++ skills/meet-join/tools/__tests__/meet-leave-tool.test.ts
@@ -27,7 +27,7 @@ const getSessionMock = mock((meetingId: string) => {
   return found ?? null;
 });
 
-mock.module("../../../../../skills/meet-join/daemon/session-manager.js", () => ({
+mock.module("../../daemon/session-manager.js", () => ({
   MeetSessionManager: {
     join: async () => {
       throw new Error("join should not be invoked in leave tests");
@@ -38,31 +38,33 @@ mock.module("../../../../../skills/meet-join/daemon/session-manager.js", () => (
   },
 }));
 
-mock.module("../../../config/assistant-feature-flags.js", () => ({
-  isAssistantFeatureFlagEnabled: (key: string) => {
-    if (key === "meet") return flagEnabled;
-    return true;
-  },
-}));
+mock.module(
+  "../../../../assistant/src/config/assistant-feature-flags.js",
+  () => ({
+    isAssistantFeatureFlagEnabled: (key: string) => {
+      if (key === "meet") return flagEnabled;
+      return true;
+    },
+  }),
+);
 
-mock.module("../../../config/loader.js", () => ({
+mock.module("../../../../assistant/src/config/loader.js", () => ({
   getConfig: () => ({
     services: { meet: { consentMessage: "unused-in-leave-tests" } },
   }),
 }));
 
-mock.module("../../../util/logger.js", () => ({
+mock.module("../../../../assistant/src/util/logger.js", () => ({
   getLogger: () =>
     new Proxy({} as Record<string, unknown>, {
       get: () => () => {},
     }),
 }));
 
-const { meetLeaveTool, DEFAULT_LEAVE_REASON } = await import(
-  "../meet-leave-tool.js"
-);
+const { meetLeaveTool, DEFAULT_LEAVE_REASON } =
+  await import("../meet-leave-tool.js");
 
-import type { ToolContext } from "../../types.js";
+import type { ToolContext } from "../../../../assistant/src/tools/types.js";
 
 function makeContext(overrides: Partial<ToolContext> = {}): ToolContext {
   return {
